Moved some types around in preparation for new logic
